Can add observations. Can change event, and subsequently, activity, manually.
The user has to select an event when they launch the app for the first time. The user must also select an activity from the list of activities for the event they selected. The process is the same when the user changes the event manually. The user is able to search through all the subjects and add an observation for that subject linking the subject event and activity together along with a recording of the score/value that the subect got doing the activity.